Move remaining analyzers to common analysis module
This commit moves all the remaining analyzers except the standard analyzer from server to the common analysis module.

Relates to #23658
